requirements.txt
allure-pytest==2.14.1
allure-python-commons==2.14.1
apispec==6.8.1
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Babel==2.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
greenlet==3.2.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pluggy==1.5.0
prison==0.2.1
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rich==13.9.4
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
